Add/ 오늘의 픽 창 UI
