+++ src/assets/color.js
@@ -6,3 +6,4 @@ export const COLOR_TEXT_DARKGRAY = '#535353';
 export const COLOR_GRAY = '#888888';
 export const COLOR_WHITE = '#FFFFFF';
 export const COLOR_TEXT70GRAY = '#4D4D4D';
+export const COLOR_TEXT60GRAY = '#666666';
