+++ src/screens/home/HomeScreen.js
@@ -34,14 +34,13 @@ import AppContext from '../../components/AppContext';
 import axios from 'axios';
 import {API_URL} from '@env';
 import {Dimensions} from 'react-native';
+import TodayPick from '../../components/TodayPick';
 
 const windowWidth = Dimensions.get('window').width;
 
 export default function HomeScreen() {
   const navigation = useNavigation();
   const context = useContext(AppContext);
-
-  const scrollViewRef = useRef();
 
   const [todaysPick, setTodaysPick] = useState([
     {
@@ -51,6 +50,10 @@ export default function HomeScreen() {
       image: 'https://d2da4yi19up8sp.cloudfront.net/product/max.jpeg',
       score: 4.5,
       reviewCount: 100,
+      firstReview: {
+        reviewer: '엄승주',
+        body: '가게 내부 깨끗하고, 서비스 친절해서 개좋음. 기대안하고 첨 갔는데 걍 인생 oo 맛봄. 지림.',
+      },
     },
     {
       name: '무대뽀 핫도그',
@@ -59,6 +62,10 @@ export default function HomeScreen() {
       image: 'https://d2da4yi19up8sp.cloudfront.net/product/pro.jpeg',
       score: 4.5,
       reviewCount: 100,
+      firstReview: {
+        reviewer: '엄승주',
+        body: '가게 내부 깨끗하고, 서비스 친절해서 개좋음. 기대안하고 첨 갔는데 걍 인생 oo 맛봄. 지림.',
+      },
     },
     {
       name: '무대뽀 핫도그',
@@ -67,118 +74,20 @@ export default function HomeScreen() {
       image: 'https://d2da4yi19up8sp.cloudfront.net/product/pro.jpeg',
       score: 4.5,
       reviewCount: 100,
+      firstReview: {
+        reviewer: '엄승주',
+        body: '가게 내부 깨끗하고, 서비스 친절해서 개좋음. 기대안하고 첨 갔는데 걍 인생 oo 맛봄. 지림.',
+      },
     },
   ]);
-
-  // const handleScroll = event => {
-  //   const offsetX = event.nativeEvent.contentOffset.x;
-  //   // console.log('offsetX', event.nativeEvent.layoutMeasurement.width);
-  //   const pageWidth = event.nativeEvent.layoutMeasurement.width;
-  //   const currentPage = Math.floor(offsetX / pageWidth + 0.5);
-  //   // setCurrentPage(currentPage);
-  //   // setCurrentDetailText(currentPage);
-  // };
 
   return (
     <>
       <Header title={'홈'} isBackButton={false} />
-      <View style={styles.entire}>
+      <ScrollView contentContainerStyle={styles.entire}>
         {/* 먹구스꾸 오늘의 픽 */}
-        <View style={styles.todayPick}>
-          <Text style={styles.todayPickTitle}>먹구스꾸's 오늘의 픽</Text>
-
-          <ScrollView
-            showsVerticalScrollIndicator={false}
-            showsHorizontalScrollIndicator={true}
-            style={{
-              // backgroundColor: 'green',
-              marginTop: 10,
-            }}
-            horizontal={true}
-            pagingEnabled
-            // onScroll={handleScroll}
-            scrollEventThrottle={16}
-            ref={scrollViewRef}>
-            {todaysPick.map((pickData, index) => {
-              return (
-                <Pressable
-                  key={index.toString()}
-                  style={{
-                    width: windowWidth - 52,
-                  }}
-                  onPress={() => {
-                    console.log('가게 상세 페이지로 이동');
-                  }}>
-                  <View style={{flexDirection: 'row'}}>
-                    <Image
-                      source={{
-                        uri: pickData.image,
-                      }}
-                      resizeMode="cover"
-                      style={{
-                        width: windowWidth / 3,
-                        height: windowWidth / 3,
-                        borderRadius: 16,
-                      }}
-                    />
-                    <View
-                      style={{
-                        flex: 1,
-                        marginLeft: 12,
-                        // backgroundColor: 'green',
-                      }}>
-                      <View
-                        style={{flexDirection: 'row', alignItems: 'center'}}>
-                        <Text
-                          style={{
-                            fontSize: 17,
-                            color: COLOR_TEXT_BLACK,
-                            fontWeight: 'bold',
-                            alignSelf: 'center',
-                          }}>
-                          {pickData.name}
-                        </Text>
-                        <SvgXml
-                          xml={svgXml.icon.star}
-                          width="15"
-                          height="15"
-                          style={{marginLeft: 7}}
-                        />
-                        <Text
-                          style={{
-                            fontSize: 12,
-                            color: COLOR_TEXT70GRAY,
-                            fontWeight: 'bold',
-                            alignSelf: 'center',
-                            marginLeft: 7,
-                          }}>
-                          {pickData.score + ' (' + pickData.reviewCount + ')'}
-                        </Text>
-                      </View>
-
-                      <Text
-                        style={{
-                          fontSize: 13,
-                          color: COLOR_GRAY,
-                        }}>
-                        {pickData.category}
-                      </Text>
-
-                      <Text
-                        style={{
-                          fontSize: 13,
-                          color: COLOR_TEXT_BLACK,
-                        }}>
-                        {pickData.menu}
-                      </Text>
-                    </View>
-                  </View>
-                </Pressable>
-              );
-            })}
-          </ScrollView>
-        </View>
-      </View>
+        <TodayPick todaysPick={todaysPick} />
+      </ScrollView>
     </>
   );
 }
@@ -187,18 +96,7 @@ const styles = StyleSheet.create({
   entire: {
     flex: 1,
     backgroundColor: COLOR_BACKGROUND,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  textMain: {
-    fontSize: 50,
-    color: COLOR_PRIMARY,
-  },
-  buttonTest: {
-    backgroundColor: COLOR_PRIMARY,
-    padding: 10,
-    borderRadius: 5,
-    justifyContent: 'center',
+    // justifyContent: 'center',
     alignItems: 'center',
   },
   buttonText: {
@@ -206,24 +104,5 @@ const styles = StyleSheet.create({
     color: COLOR_WHITE,
     fontWeight: 'bold',
     alignSelf: 'center',
-  },
-  todayPick: {
-    width: windowWidth - 32,
-    padding: 12,
-    paddingHorizontal: 10,
-    backgroundColor: COLOR_WHITE,
-    borderRadius: 10,
-    shadowOffset: {
-      width: 0,
-      height: 4,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 4,
-    elevation: 4, // for Android
-  },
-  todayPickTitle: {
-    fontSize: 20,
-    color: COLOR_TEXT70GRAY,
-    fontWeight: '700',
   },
 });
